+++ eoreader/products/optical/re_product.py
@@ -24,14 +24,13 @@ for more information.
 import logging
 from datetime import datetime
 from enum import unique
-from pathlib import Path
 from typing import Union
 
 import numpy as np
 import xarray as xr
-from cloudpathlib import CloudPath
 from sertit import files, path
 from sertit.misc import ListEnum
+from sertit.types import AnyPathType
 
 from eoreader import DATETIME_FMT, EOREADER_NAME
 from eoreader.bands import (
@@ -335,7 +334,7 @@ class ReProduct(PlanetProduct):
     def _to_reflectance(
         self,
         band_arr: xr.DataArray,
-        band_path: Union[Path, CloudPath],
+        band_path: AnyPathType,
         band: BandNames,
         **kwargs,
     ) -> xr.DataArray:
@@ -344,7 +343,7 @@ class ReProduct(PlanetProduct):
 
         Args:
             band_arr (xr.DataArray): Band array to convert
-            band_path (Union[CloudPath, Path]): Band path
+            band_path (AnyPathType): Band path
             band (BandNames): Band to read
             **kwargs: Other keywords