+++ eoreader/products/optical/spot45_product.py
@@ -22,16 +22,14 @@ for more information.
 import logging
 from datetime import timedelta
 from enum import unique
-from pathlib import Path
-from typing import Union
 
 import numpy as np
 import xarray as xr
-from cloudpathlib import CloudPath
 from lxml import etree
 from rasterio import crs as riocrs
 from sertit import path
 from sertit.misc import ListEnum
+from sertit.types import AnyPathType
 
 from eoreader import DATETIME_FMT, EOREADER_NAME, cache
 from eoreader.bands import (
@@ -544,7 +542,7 @@ class Spot45Product(DimapV1Product):
     def _to_reflectance(
         self,
         band_arr: xr.DataArray,
-        band_path: Union[Path, CloudPath],
+        band_path: AnyPathType,
         band: BandNames,
         **kwargs,
     ) -> xr.DataArray:
@@ -553,7 +551,7 @@ class Spot45Product(DimapV1Product):
 
         Args:
             band_arr (xr.DataArray): Band array to convert
-            band_path (Union[CloudPath, Path]): Band path
+            band_path (AnyPathType): Band path
             band (BandNames): Band to read
             **kwargs: Other keywords
 
@@ -608,12 +606,12 @@ class Spot45Product(DimapV1Product):
 
         return self._read_mtd_xml(mtd_from_path, mtd_archived)
 
-    def _get_tile_path(self) -> Union[CloudPath, Path]:
+    def _get_tile_path(self) -> AnyPathType:
         """
         Get the DIMAP filepath
 
         Returns:
-            Union[CloudPath, Path]: DIMAP filepath
+            AnyPathType: DIMAP filepath
         """
         return self._get_path("IMAGERY", "TIF")